Added edit modal functionality
